Hardware fault exception division by zero fixed (#166)
* created a new branch based on commit dd453fc to test hardware fault handling

In this branch motor is working if division by zero exception is not enabled.
Before on the old branch(based on PR 158) nothing was working.
So, a new branch was created from the PR 158 and tested on the bike.
Bike doesn't works using the PR 158. Because of this it was created this  branch based on the last commit on the main.

* Fixed some division by zero detect by the hardware_falut_handle

Tests with the hardware_falut_handle was cause the system reset because a division by was detected.
The division by zero problem was fixed in void LCD_APT_frame_Process(APT_Handle_t *pHandle) and WheelSpdSensor_GetSpeedFreq.

* implemented recommendations from Andy and Francois.

This recommendations come from the PR 166.