Fixed small bug from a conflict on the PR173/PR168
PR173 was using one version of the main before PR168.PR168 was merger and one modification on one define and one variable, used by the function created on the PR173, was done. After the merger theses variables were not defined anymore and this created a error one some functions created in the PR173.