Grizzly vibration fix (#268)
* conditional KPI

only for VELEC, the KPI parameters are changing when the throttle is released.

* Delete RTX_Config.c.base@5.1.1

* Delete RTX_Config.h.base@5.5.2

* Delete startup_RA.c

* Delete startup_RA.c.base@1.0.0

* Delete system_RA.c

* Delete system_RA.c.base@1.0.0

* Delete startup_RA.c.base@1.0.0

* Delete system_RA.c.base@1.0.0

* comment added

* maximizing torque and power

the max. torque was increased to 1500 and max power was set to 1500W

* Revert "Merge branch 'main' of https://github.com/tristar-multicopters/GNR_firmware"

This reverts commit d75e7e30d844d6e7d84c2a98ba9c06395b26cef0, reversing
changes made to 9d386becc5ad6a9925ccb5b81e6560e443580c2e.

* vibration_fix_forGrizzly

the grizzly bike was tuned with the Autotuune values but , it seems they are not good for high speed. then I changes the high speed set of Kp,Ki with the old values.

---------

Co-authored-by: BehnamShakibafar <behnam.shakibafar@ftex.ca>
Co-authored-by: Behnam Shakibafar <115581475+BehnamShakibafar@users.noreply.github.com>