Increase the maximum J parameter (motor inertia) value in autotune. (#649)
* Increase the maximum J value in autotune.

When I tried to do autotune for the Super73 1200W, I always got a J_error. After investigating the autotune library, I discovered that there is a default maximum J value, and the J value for that motor exceeded it. To prevent this error for other motors we may have in the future, I increased the maximum J value by ten times.

* use constants for minimum and maximum j value