Change motor temperature configuraiton (#685)
* remove ntc table file and add calculate temperature functions

remove ntc table file and all their dependencies and add calculate temperature functions base on coefficient, however, this commit needs clean-ups.

* create hardware parameters

* add ntc source parameter

add ntc source parameter to to know the data is from motor or heatsink.
and indentation.

* fix names of hardware parameter

* remove ntc outsidetable parameter

* change byte to tics

* correct adc parameter conversion

* change rated resistance to resistance coef

I understand that the problem of my PR was in the exp() function (I don't know why, maybe due to an overflow issue or a big loop issue). So, I changed the rated resistance parameter to another coefficient, which should be calculated based on the rated resistance with this formula: exp(NTCBetaCoef / TEMP_25_CELSIUS_IN_KELVIN) / NTCRatedResistance. I also added this formula in the comment in front of the parameter.

* remove line

* fix conflict