PAS Deceleration Ramps (#691)
* Tuning Torque Sensor Error Issue Function

After previous changes on the power delivery mechanism and PAS detection,
I tuned the torque sensor error function for better responsiveness.

* Added Test Ramp Function

Added a test function to test the functionnality of the decel ramp

* Updated PAS Tuning Params

* Removed Ramp for throttle

On main, ramp was not changing anything on the throttle behaviour.
Will be reimplemented with the new ramps system.

* Removed PAS Tuning Header Files

Requests were made to remove the PAS tuning header files in order to have all
parameters in the vc params header files so integration could be done in 1 single file
with all information needed to do a complete integration work on a bike.

* Added Float km/h Speed Function

In order to improve accuracy for ramps/PAS detection, I have refactored
a km/h float speed into a wheel function.

* Removed Previous Ramp Implementation

After talking with the team, we decided to remove the previous ramps implementation as it
was not used currently on main. We are moving to a more modular approach, with easier
ways on adding more ramps/filtering effects on the power delivery.

* Ramps Implementation Completed

I have implemented two differents deceleration ramps. Both of these ramps
can be selected from the vc_parameters_bike header files.

* Changing Float Define Without Casts

Removed the float cast to decimal defines in order to remove the waste of computing
on casting op