Scaled Torque Gain By Pedaling RPM (#705)
Last feature of our PAS 2024 Rework. This features introduces the
scaling of our torque delivery according to the pedaling RPM. Each bike
param header files has been updated with parameters to enable/disable
this feature with also parameters to tune the scaling. For now, the
torque scaling is done before the ramp filtering.