Separate WSS from PAS source code (#711)
* Initialize WSS modular structure

* Separate WSS and PAS Source Code

This PR separates the WSS from any other relationship with any part of the VC layer code.
The WSS is not completly separated from the pVCI Handle, removing any dependencies and chain pointer link
between all the different layers of VC.
This means, wherever VC needs a value from the WSS, VC will call a get function, without accessing the reference to the structure directly.

Currently, I am planning of moving with this approach for the next 2 PAS sensors. This approach will be revisited if the team
feels the need to modify my current approach. This effort was made in order to break the ever-lasting chain of pointer reference within the VC Handle.

Note that any new necessity of access a different WSS value will need an additional getter in order to access the needed value.

* Uniform Naming Convention