PAS AND Fix (#732)
This changes removed all the other PAS mechanisms that was preventing the current AND/OR PAS mechanism to work properly.
This change was needed in order to deliver the e-cells firmware, as a AND PAS configuration is absolutely necessary for safety reasons.

Important note : My next task consist in cleaning up the previous PAS code. Considering that, at this state, the current firmware will not used any
of the previous PAS code, it is deadcode for now. This deadcode will be removed as soon as the next PR will be ready but for now, this is an important fix
in order to deliver the e-cells firmware.