Fix PAS Default Settings. (#733)
A few changes were made in the commit.
The first one being added a new PAS parameter check to detect cadence,
instead of using the previous PAS parameters. This allows better responsiveness and
cohesion while using the new PAS.

The second change is a update on the default PAS parameters in the user config.
This also fixes a lot of issues reported here : https://tristarmulticopters.atlassian.net/browse/DEV-1282

The last changes is the addition of all possible PAS parameters to the e-vc_parameters_e_cells.h file.